update close #193 1.14.0[2.40] 第-10.7周 小寒/大寒2-6

+++ Assets/Scripts/App/StartUp.cs
@@ -17,6 +17,7 @@ namespace App
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
         private static void DebugSetting()
         {
+            if (StartUp.self == null) return;
             if (StartUp.self.clearDir && Debug.isDebugBuild && !Application.isEditor &&
                 PlayerPrefs.GetString(Loading.VersionKey) != Application.version)
                 Directory.Delete(Application.persistentDataPath + "/com.unity.addressables", true);
